+++ amorphie.consent.data/Helper/FullTextSearchHelper.cs
